+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Added documentation on the aggregated-translations tool
+
 
 3.0.1 - (April 24, 2018)
 ----------